refactor: RouteConverter 삭제 후 좌표 변환 책임을 Piece에게 위임
- 좌표 변환 메서드를 piece에 만들어놨지만 PieceType에서 처리
