+++ src/main/java/janggi/domain/piece/Piece.java
@@ -1,8 +1,11 @@
 package janggi.domain.piece;
 
+import janggi.domain.board.Board;
+import janggi.domain.common.Position;
 import janggi.domain.common.Team;
 import janggi.domain.route.Route;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 
 public class Piece {
@@ -41,6 +44,10 @@ public class Piece {
 
     public boolean isSameTeam(Piece other) {
         return team == other.team;
+    }
+
+    public Map<Position, List<Position>> convertToPosition(Board board, Position position) {
+        return pieceType.convertToPosition(board, position);
     }
 
     @Override
